Update VirtualModulesPlugin from upstream
This updates the internal VirtualModulesPlugin, which is copied from
https://github.com/sysgears/webpack-virtual-modules. I went through all
commits since this virtual.js was added, and applied them manually to
the svelte-loader version where applicable. These upstream changes
allow for using Webpack 5 with svelte-loader. I have also merged in the
changes of this patch, since it fixes :
https://github.com/sysgears/webpack-virtual-modules/commit/0c8926e0a82a4764f8f53133da8b903a3558f171

This also adds a LICENSE file to comply with the requirements of the MIT
license for the plugin files. The LICENSE file also adds in the text of
MIT license that svelte-loader is licensed under (with a different
copyright line, since svelte-loader wasn't written by SysGears)